requirements.txt
anthropic==0.18.1
openai==1.12.0
python-dotenv==1.0.1
flask==3.0.2 